requirements.txt
python-dotenv
requests
sp-api==0.19.7
